faker as shared package between Micro Frontend
