+++ ecommerce/cart/src/index.js
@@ -1,5 +1,1 @@
-import faker from 'faker';
-
-const message = `<div>You have ${faker.random.number()} items in your cart</div>`;
-
-document.querySelector('#dev-cart').innerHTML = message;
+import('./boostrap');
